+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   **Infra** Added Better Uptime monitor
 -   **Bolt** Add Docker `RUN` cache to distributed deploys to improve deploy speeds
+-   **Infra** Prometheus VPA
 
 ### Changed